FIX: item types import error
